+++ internal/infrastructure/datasource/entity/processed_signal_entity.go
@@ -11,6 +11,9 @@ type ProcessedSignalEntity struct {
 	Wavelength        float64                `gorm:"not null"`
 	Polarization      string                 `gorm:"size:10;not null"`
 	IsPhoton          bool                   `gorm:"not null"`
+	DeviceID          string                 `gorm:"size:255;default:''"`
+	BinWidth          float64                `gorm:"default:0"`
+	NDataPoints       int                    `gorm:"default:0"`
 	Signal            datatypes.Float64Slice `gorm:"type:bytea;not null"`
 }